refactor: rename deleteDouble to deleteDuplicates
